api connected to SQL

+++ RecipeAPI/Models/Recipe.cs
@@ -5,7 +5,7 @@
         public Guid Id { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
-        public double TimeToMake { get; set; }
+        public decimal TimeToMake { get; set; }
         public string ImageUrl { get; set; }
         public List<string> Ingredients { get; set; }
         public List<string> Instructions { get; set; }
